requirements.txt
setuptools
selenium
python-dotenv
chromedriver-autoinstaller